# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CaterOS SaaS API Documentation",
        "description": "Dokumentasi RESTful API untuk platform SaaS Management Catering End-to-End CaterOS.",
        "contact": {
            "name": "CaterOS Engineering Team",
            "email": "dev@cateros.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/tenant/menu-categories": {
            "get": {
                "tags": [
                    "Master Kategori Menu"
                ],
                "summary": "Daftar Kategori Menu",
                "description": "Get all menu categories for active tenant.",
                "operationId": "6207379297a8895c724f3808c5ebb6d7",
                "responses": {
                    "200": {
                        "description": "Daftar kategori berhasil diambil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "TenantHeader": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Master Kategori Menu"
                ],
                "summary": "Tambah Kategori Menu Baru",
                "description": "Store new menu category.",
                "operationId": "44f3e247034c339020ab747597feac2b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Lauk Utama Ayam & Bebek"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Aneka olahan lauk ayam dan bebek bakar/goreng"
                                    },
                                    "sort_order": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Kategori menu berhasil dibuat"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "TenantHeader": []
                    }
                ]
            }
        },
        "/tenant/menu-categories/{id}": {
            "put": {
                "tags": [
                    "Master Kategori Menu"
                ],
                "summary": "Update Kategori Menu",
                "description": "Update menu category.",
                "operationId": "704e7a9eaae398eed41881b78b528416",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Lauk Utama Ayam & Unggas"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Update deskripsi kategori"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kategori berhasil diperbarui"
                    },
                    "404": {
                        "description": "Kategori tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "TenantHeader": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Master Kategori Menu"
                ],
                "summary": "Hapus Kategori Menu",
                "description": "Delete menu category.",
                "operationId": "79ee55f1dcaed0c7f434dbb1d352c3c7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kategori berhasil dihapus"
                    },
                    "400": {
                        "description": "Kategori masih berisi menu item"
                    },
                    "404": {
                        "description": "Kategori tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "TenantHeader": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan Token Bearer Sanctum (contoh: 1|abcdef...)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "TenantHeader": {
                "type": "apiKey",
                "description": "Header ID Tenant yang sedang aktif (contoh: 1)",
                "name": "X-Tenant-ID",
                "in": "header"
            }
        }
    },
...
}